More descriptive filenames. Stop relying on folder names.
Life is easier as a developer when you can quickly identify a file by
its filename. Currently the file paths are informative, but if you
only see the filename, it's difficult to work with.
