+++ tests/extended_temporal_memory/etm_sequence_memory_cpp_test.py
@@ -1,6 +1,6 @@
 # ----------------------------------------------------------------------
 # Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2017, Numenta, Inc.  Unless you have an agreement
+# Copyright (C) 2016-2017, Numenta, Inc.  Unless you have an agreement
 # with Numenta, Inc., for a separate license for this software code, the
 # following terms and conditions apply:
 #
@@ -19,24 +19,20 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-import operator
 import unittest
 
-import numpy as np
+from nupic.bindings.experimental import ExtendedTemporalMemory
+from htmresearch.support.shared_tests.sequence_memory_test_base import (
+  SequenceMemoryTestBase)
 
-from htmresearch.algorithms.sparsematrix_temporal_memory.basal_context_apical_disambiguation import TemporalMemory
-from htmresearch.support.temporal_memory_tests.sequence_memory import SequenceMemoryTestBase
 
-
-class ApicalTiebreakTM_SequenceMemoryTests(SequenceMemoryTestBase,
-                                           unittest.TestCase):
+class ExtendedTemporalMemoryCPP_SequenceMemoryTests(SequenceMemoryTestBase,
+                                                    unittest.TestCase):
 
   def constructTM(self, columnDimensions, cellsPerColumn, initialPermanence,
                   connectedPermanence, minThreshold, sampleSize,
                   permanenceIncrement, permanenceDecrement,
                   predictedSegmentDecrement, activationThreshold, seed):
-
-    numColumns = reduce(operator.mul, columnDimensions, 1)
 
     params = {
       "columnDimensions": columnDimensions,
@@ -44,26 +40,23 @@ class ApicalTiebreakTM_SequenceMemoryTests(SequenceMemoryTestBase,
       "initialPermanence": initialPermanence,
       "connectedPermanence": connectedPermanence,
       "minThreshold": minThreshold,
-      "sampleSize": sampleSize,
+      "maxNewSynapseCount": sampleSize,
       "permanenceIncrement": permanenceIncrement,
       "permanenceDecrement": permanenceDecrement,
       "predictedSegmentDecrement": predictedSegmentDecrement,
       "activationThreshold": activationThreshold,
       "seed": seed,
-      "basalInputDimensions": (numColumns*cellsPerColumn,),
-      "apicalInputDimensions": (),
+      "learnOnOneCell": False,
     }
 
-    self.tm = TemporalMemory(**params)
+    self.tm = ExtendedTemporalMemory(**params)
 
 
   def compute(self, activeColumns, learn):
-    activeColumns = np.array(sorted(activeColumns), dtype="uint32")
-
-    self.tm.compute(activeColumns,
-                    basalInput=self.tm.getActiveCells(),
-                    basalGrowthCandidates=self.tm.getWinnerCells(),
-                    learn=learn)
+    # Use depolarizeCells + activateCells rather than tm.compute so that
+    # getPredictiveCells returns predictions for the current timestep.
+    self.tm.depolarizeCells(learn=learn)
+    self.tm.activateCells(sorted(activeColumns), learn=learn)
 
 
   def reset(self):
@@ -75,4 +68,4 @@ class ApicalTiebreakTM_SequenceMemoryTests(SequenceMemoryTestBase,
 
 
   def getPreviouslyPredictedCells(self):
-    return self.tm.getPreviouslyPredictedCells()
+    return self.tm.getPredictiveCells()
